up bbb to my fork

+++ record-and-playback/presentation/scripts/process/presentation.rb
@@ -42,11 +42,13 @@ raw_archive_dir = "#{recording_dir}/raw/#{meeting_id}"
 target_dir = "#{recording_dir}/process/presentation/#{meeting_id}"
 if not FileTest.directory?(target_dir)
   FileUtils.mkdir_p "/var/log/bigbluebutton/presentation"
-	logger = Logger.new("/var/log/bigbluebutton/presentation/process-#{meeting_id}.log", 'daily' )
-	BigBlueButton.logger = logger
+  logger = Logger.new("/var/log/bigbluebutton/presentation/process-#{meeting_id}.log", 'daily' )
+  BigBlueButton.logger = logger
   BigBlueButton.logger.info("Processing script presentation.rb")
-	FileUtils.mkdir_p target_dir
+  FileUtils.mkdir_p target_dir
   
+ begin
+
   # Create a copy of the raw archives
   temp_dir = "#{target_dir}/temp"
   FileUtils.mkdir_p temp_dir
@@ -65,7 +67,6 @@ if not FileTest.directory?(target_dir)
   presentations.each do |pres|
     pres_dir = "#{presentation_dir}/#{pres}"
     num_pages = BigBlueButton::Presentation.get_number_of_pages_for(pres_dir)
-    pres_pdf = "#{pres_dir}/#{pres}.pdf"
     
     target_pres_dir = "#{processed_pres_dir}/#{pres}"
     FileUtils.mkdir_p target_pres_dir
@@ -73,29 +74,35 @@ if not FileTest.directory?(target_dir)
     
     images=Dir.glob("#{pres_dir}/#{pres}.{jpg,png,gif,JPG,PNG,GIF}")
     if images.empty? 
-         1.upto(num_pages) do |page| 
-           pdf_page = "#{pres_dir}/slide-#{page}.pdf"
-           BigBlueButton::Presentation.extract_page_from_pdf(page, pres_pdf, pdf_page)
-           #BigBlueButton::Presentation.convert_pdf_to_png(pdf_page, "#{target_pres_dir}/slide-#{page}.png")
-           command = "convert -density 300x300 -resize 1600x1200 -quality 90 +dither -depth 8 -colors 256 #{pdf_page} #{target_pres_dir}/slide-#{page}.png"
-           BigBlueButton.execute(command)
-           if File.exist?("#{pres_dir}/textfiles/slide-#{page}.txt") then
-             FileUtils.cp("#{pres_dir}/textfiles/slide-#{page}.txt", "#{target_pres_dir}/textfiles")
-           end
-         end
-    else
-        ext = File.extname("#{images[0]}")
-        #BigBlueButton::Presentation.convert_image_to_png(images[0],"#{target_pres_dir}/slide-1.png")
-        command="convert -resize 1600x1200 #{images[0]} #{target_pres_dir}/slide-1.png"
+      pres_pdf = "#{pres_dir}/#{pres}.pdf"
+      if !File.exists?(pres_pdf)
+        BigBlueButton.logger.info("Falling back to old presentation filename")
+        pres_pdf = "#{pres_dir}/#{pres}"
+      end
+      if !File.exists?(pres_pdf)
+        raise "Could not find pdf file for presentation #{pres}"
+      end
+      1.upto(num_pages) do |page| 
+        pdf_page = "#{pres_dir}/slide-#{page}.pdf"
+        BigBlueButton::Presentation.extract_page_from_pdf(page, pres_pdf, pdf_page)
+        #BigBlueButton::Presentation.convert_pdf_to_png(pdf_page, "#{target_pres_dir}/slide-#{page}.png")
+        command = "convert -density 300x300 -resize 1600x1200 -quality 90 +dither -depth 8 -colors 256 #{pdf_page} #{target_pres_dir}/slide-#{page}.png"
         BigBlueButton.execute(command)
+        if File.exist?("#{pres_dir}/textfiles/slide-#{page}.txt") then
+          FileUtils.cp("#{pres_dir}/textfiles/slide-#{page}.txt", "#{target_pres_dir}/textfiles")
+        end
+      end
+    else
+      ext = File.extname("#{images[0]}")
+      #BigBlueButton::Presentation.convert_image_to_png(images[0],"#{target_pres_dir}/slide-1.png")
+      command="convert -resize 1600x1200 #{images[0]} #{target_pres_dir}/slide-1.png"
+      BigBlueButton.execute(command)
     end
   
   end
   
   if !Dir["#{raw_archive_dir}/video/*"].empty?    
     BigBlueButton.process_multiple_videos(target_dir, temp_dir, meeting_id, presentation_props['video_output_width'], presentation_props['video_output_height'])
-  else    
-    BigBlueButton::AudioProcessor.process("#{temp_dir}/#{meeting_id}", "#{target_dir}/audio.ogg")
   end
 
   process_done = File.new("#{recording_dir}/status/processed/#{meeting_id}-presentation.done", "w")
@@ -103,5 +110,8 @@ if not FileTest.directory?(target_dir)
   process_done.close
 #else
 #	BigBlueButton.logger.debug("Skipping #{meeting_id} as it has already been processed.")  
+ rescue Exception => e
+        BigBlueButton.logger.error(e.message)
+ end
 end
